refactor(server): unify LoomServer with version-specific registration proof
