+++ build.mill
@@ -23,6 +23,10 @@ object Versions {
 val scalaVersions = Seq(Versions.Scala213, Versions.Scala3)
 val wsRoot        = BuildCtx.workspaceRoot
 
+/** Version-specific source dir for cross-built modules (`scala-2` on 2.x, `scala-3` on 3.x). */
+def vDir(crossScalaVersion: String): String =
+  if (crossScalaVersion.startsWith("2.")) "scala-2" else "scala-3"
+
 trait ZioHttpPublishModule extends PublishModule {
   def publishVersion = Task.Input {
     sys.env.get("ZIO_HTTP_SNAPSHOT_VERSION").getOrElse {
@@ -372,15 +376,8 @@ object serverLoom extends Module {
   trait JvmModule extends ZioHttpJvmModule {
     def artifactName     = "zio-http-server-loom"
     def moduleDeps       = Seq(server.jvm(), h2Codec.jvm())
-    override def sources = {
-      // LoomServer is version-split: the typed tuple-keyed builder
-      // (NotGiven/Tuple.Contains) is Scala 3 only, with a Scala 2 fallback
-      // keeping the same runtime behavior (same pattern as core/endpoint).
-      val vDir =
-        if (crossScalaVersion.startsWith("2.")) "scala-2" else if (crossScalaVersion.startsWith("3.")) "scala-3" else ""
-      val base = Seq(r / "src" / "main" / "scala")
-      Task.Sources((if (vDir.nonEmpty) base :+ (r / "src" / "main" / vDir) else base) *)
-    }
+    // Shared sources plus the Scala 2/3 engine-proof variant.
+    override def sources = Task.Sources(r / "src" / "main" / "scala", r / "src" / "main" / vDir(crossScalaVersion))
     // Branch minimum is 75%: ~25% of branches are in private defensive Any-match arms
     // (toStream, toUnit, toOptionalFrame, toReceivedFrame, toSendError) that exist for
     // Mux API version compatibility and cannot be reached from outside the enclosing class.
@@ -393,15 +390,8 @@ object serverLoom extends Module {
     )
     object test extends ScoverageTests with ZioHttpTestModule {
       override def moduleDeps = Seq(serverLoom.jvm(), clientJava.jvm())
-      override def sources    = {
-        // The typed-negation spec uses Scala 3 compile-time testing only.
-        val vDir =
-          if (crossScalaVersion.startsWith("2.")) "scala-2"
-          else if (crossScalaVersion.startsWith("3.")) "scala-3"
-          else ""
-        val base = Seq(r / "src" / "test" / "scala")
-        Task.Sources((if (vDir.nonEmpty) base :+ (r / "src" / "test" / vDir) else base) *)
-      }
+      // Shared specs plus the Scala 3 compile-time negation spec.
+      override def sources    = Task.Sources(r / "src" / "test" / "scala", r / "src" / "test" / vDir(crossScalaVersion))
     }
   }
 }
